+++ debug_utils.py
@@ -1,5 +1,5 @@
 import os
-from colors import pr_light_purple,pr_yellow,pr_pink,pr_green
+from colors import pr_light_purple,pr_yellow,pr_pink,pr_green,pr_red
 from constants import fields_dict
 
 VERBOSE = os.getenv('VERBOSE')
@@ -9,7 +9,8 @@ def debug_steps(row,msg,level):
         LOG = f"[DEBUG] - Level {level} - {msg}"
         current_message = row[fields_dict[level]]
         row[fields_dict[level]] = f"{current_message}\n{LOG}"
-        pr_green(LOG) 
+        pr_red(LOG) 
+        print("="*20)
         
 def debug(msg):
     if VERBOSE=="True":